- 안쓰는 JDBC Repo 쿼리 삭제 - 최신 게시글 조회 기능 JPARepo 사용하도록 수정

+++ MyBoard/src/main/java/com/ltj/myboard/dto/post/FilteredPost.java
@@ -13,6 +13,6 @@ import lombok.ToString;
 @ToString
 public class FilteredPost {
     private int OrderedPostNo;
-    private int CommentCount;
+    private long CommentCount;
     private Post PostData;
 }
